Refactor finansiering => studieFinansiering. Flyttet regler ett nivå opp.
