+++ src/regler/domeneRegel.js
